# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  created: "created",
  payment_pending: "payment_pending",
  payment_confirmed: "payment_confirmed",
  supplier_processing: "supplier_processing",
  fulfilled: "fulfilled",
  email_sent: "email_sent",
  payment_failed: "payment_failed",
  supplier_failed: "supplier_failed",
  email_failed: "email_failed",
  failed: "failed",
  cancelled: "cancelled",
  refunded: "refunded",
} as const;

export type CreateOrderInputCheckoutData = { [key: string]: string };

export interface CreateOrderInput {
  productSlug: string;
  variantId: string;
  email: string;
  checkoutKey: string;
  checkoutData?: CreateOrderInputCheckoutData;
}

export interface CreateOrderResult {
  publicId: string;
  accessToken: string;
  paymentUrl: string;
}

export interface PublicOrder {
  publicId: string;
  status: OrderStatus;
  productName: string;
  nominalLabel: string;
  emailMasked: string;
  notificationEligible: boolean;
  /** @nullable */
  errorMessage: string | null;
}

export type StorefrontCategoryId =
  (typeof StorefrontCategoryId)[keyof typeof StorefrontCategoryId];

export const StorefrontCategoryId = {
  apple: "apple",
  steam: "steam",
  games: "games",
  telegram: "telegram",
} as const;

export interface StorefrontCategory {
  id: StorefrontCategoryId;
  name: string;
  emoji: string;
  order: number;
  productCount: number;
}

export type StorefrontOfferNominal = {
  amount: string;
  currency: string;
};

export interface StorefrontOffer {
  id: string;
  label: string;
  nominal: StorefrontOfferNominal;
  /** @minimum 1 */
  priceRub: number;
  available: boolean;
  /** @nullable */
  stock: number | null;
}

export type SteamCurrency = (typeof SteamCurrency)[keyof typeof SteamCurrency];

export const SteamCurrency = {
  USD: "USD",
  RUB: "RUB",
  UAH: "UAH",
  KZT: "KZT",
} as const;

export interface SteamAmountRule {
  currency: SteamCurrency;
  maxFractionDigits: number;
}

export interface SteamFormConfig {
  currencies: SteamCurrency[];
  amountRules: SteamAmountRule[];
  /** @nullable */
  minimumAmount: null;
  /** @nullable */
  maximumAmount: null;
}

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  gift_card: "gift_card",
  steam_topup: "steam_topup",
  telegram_stars: "telegram_stars",
  telegram_premium: "telegram_premium",
  game_topup: "game_topup",
} as const;

export type CheckoutFieldType =
  (typeof CheckoutFieldType)[keyof typeof CheckoutFieldType];

export const CheckoutFieldType = {
  text: "text",
} as const;

export interface CheckoutField {
  key: string;
  label: string;
  type: CheckoutFieldType;
}

export interface CheckoutConfig {
  orderType: OrderType;
  supported: boolean;
  /** @nullable */
  message: string | null;
  fields: CheckoutField[];
}

export interface StorefrontProduct {
  slug: string;
  categoryId: StorefrontCategoryId;
  title: string;
  description: string;
  flag: string;
  /** @nullable */
  region: string | null;
  available: boolean;
  offers: StorefrontOffer[];
  steamForm: SteamFormConfig | null;
  checkout: CheckoutConfig;
}

export interface SteamQuoteInput {
  /**
   * @minLength 1
   * @maxLength 100
   */
  steamLogin: string;
  currency: SteamCurrency;
  /**
   * @minLength 1
   * @maxLength 32
   */
  amount: string;
}

export interface SteamQuote {
  canRefill: true;
  currency: SteamCurrency;
  amount: string;
  /** @minimum 1 */
  priceRub: number;
}

export type StorefrontCategories200 = {
  categories: StorefrontCategory[];
};

export type StorefrontProducts200 = {
  products: StorefrontProduct[];
};

export type StorefrontSteamQuote200 = {
  quote: SteamQuote;
};

export type StorefrontProduct200 = {
  product: StorefrontProduct;
};