# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: storefront
    description: Curated public storefront
  - name: orders
    description: Protected checkout and order status
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /storefront/categories:
    get:
      operationId: storefrontCategories
      tags: [storefront]
      summary: List the four public storefront categories
      responses:
        "200":
          description: Storefront categories
          content:
            application/json:
              schema:
                type: object
                required: [categories]
                properties:
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/StorefrontCategory"
  /storefront/products:
    get:
      operationId: storefrontProducts
      tags: [storefront]
      summary: List the nine curated storefront products with live offers
      responses:
        "200":
          description: Curated products
          content:
            application/json:
              schema:
                type: object
                required: [products]
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/StorefrontProduct"
  /storefront/steam/quote:
    post:
      operationId: storefrontSteamQuote
      tags: [storefront]
      summary: Validate a Steam login and calculate a read-only top-up quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SteamQuoteInput"
      responses:
        "200":
          description: Final customer quote in RUB; no order is created
          content:
            application/json:
              schema:
                type: object
                required: [quote]
                properties:
                  quote:
                    $ref: "#/components/schemas/SteamQuote"
  /storefront/products/{slug}:
    get:
      operationId: storefrontProduct
      tags: [storefront]
      summary: Get one curated storefront product
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Curated product
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product:
                    $ref: "#/components/schemas/StorefrontProduct"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an idempotent checkout using a server-verified price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Payment page created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResult"
  /orders/{publicId}:
    get:
      operationId: getOrderStatus
      tags: [orders]
      summary: Independently verify payment and return a protected order status
      security:
        - orderToken: []
      parameters:
        - $ref: "#/components/parameters/PublicOrderId"
      responses:
        "200":
          description: Safe public order state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicOrder"
  /orders/{publicId}/notification-viewed:
    post:
      operationId: markOrderNotificationViewed
      tags: [orders]
      security:
        - orderToken: []
      parameters:
        - $ref: "#/components/parameters/PublicOrderId"
      responses:
        "204":
          description: Notification acknowledged
  /orders/{publicId}/email/retry:
    post:
      operationId: retryOrderEmail
      tags: [orders]
      security:
        - orderToken: []
      parameters:
        - $ref: "#/components/parameters/PublicOrderId"
      responses:
        "202":
          description: Existing delivery result sent again without another purchase
components:
  securitySchemes:
    orderToken:
      type: http
      scheme: bearer
  parameters:
    PublicOrderId:
      in: path
      name: publicId
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    OrderStatus:
      type: string
      enum: [created, payment_pending, payment_confirmed, supplier_processing, fulfilled, email_sent, payment_failed, supplier_failed, email_failed, failed, cancelled, refunded]
    CreateOrderInput:
      type: object
      additionalProperties: false
      required: [productSlug, variantId, email, checkoutKey]
      properties:
        productSlug:
          type: string
        variantId:
          type: string
        email:
          type: string
          format: email
        checkoutKey:
          type: string
          format: uuid
        checkoutData:
          type: object
          additionalProperties:
            type: string
    CreateOrderResult:
      type: object
      required: [publicId, accessToken, paymentUrl]
      properties:
        publicId:
          type: string
        accessToken:
          type: string
        paymentUrl:
          type: string
          format: uri
    PublicOrder:
      type: object
      required: [publicId, status, productName, nominalLabel, emailMasked, notificationEligible, errorMessage]
      properties:
        publicId:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        productName:
          type: string
        nominalLabel:
          type: string
        emailMasked:
          type: string
        notificationEligible:
          type: boolean
        errorMessage:
          type: [string, "null"]
    StorefrontCategoryId:
      type: string
      enum: [apple, steam, games, telegram]
    StorefrontCategory:
      type: object
      required: [id, name, emoji, order, productCount]
      properties:
        id:
          $ref: "#/components/schemas/StorefrontCategoryId"
        name:
          type: string
        emoji:
          type: string
        order:
          type: integer
        productCount:
          type: integer
    StorefrontOffer:
      type: object
      required: [id, label, nominal, priceRub, available, stock]
      properties:
        id:
          type: string
        label:
          type: string
        nominal:
          type: object
          required: [amount, currency]
          properties:
            amount:
              type: string
            currency:
              type: string
        priceRub:
          type: integer
          minimum: 1
        available:
          type: boolean
        stock:
          type: [integer, "null"]
    StorefrontProduct:
      type: object
      required:
        [
          slug,
          categoryId,
          title,
          description,
          flag,
          region,
          available,
          offers,
          steamForm,
          checkout,
        ]
      properties:
        slug:
          type: string
        categoryId:
          $ref: "#/components/schemas/StorefrontCategoryId"
        title:
          type: string
        description:
          type: string
        flag:
          type: string
        region:
          type: [string, "null"]
        available:
          type: boolean
        offers:
          type: array
          items:
            $ref: "#/components/schemas/StorefrontOffer"
        steamForm:
          oneOf:
            - $ref: "#/components/schemas/SteamFormConfig"
            - type: "null"
        checkout:
          $ref: "#/components/schemas/CheckoutConfig"
    OrderType:
      type: string
      enum: [gift_card, steam_topup, telegram_stars, telegram_premium, game_topup]
    CheckoutField:
      type: object
      required: [key, label, type]
      properties:
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text]
    CheckoutConfig:
      type: object
      required: [orderType, supported, message, fields]
      properties:
        orderType:
          $ref: "#/components/schemas/OrderType"
        supported:
          type: boolean
        message:
          type: [string, "null"]
        fields:
          type: array
          items:
            $ref: "#/components/schemas/CheckoutField"
    SteamCurrency:
      type: string
      enum: [USD, RUB, UAH, KZT]
    SteamAmountRule:
      type: object
      required: [currency, maxFractionDigits]
      properties:
        currency:
          $ref: "#/components/schemas/SteamCurrency"
        maxFractionDigits:
          type: integer
    SteamFormConfig:
      type: object
      required: [currencies, amountRules, minimumAmount, maximumAmount]
      properties:
        currencies:
          type: array
          items:
            $ref: "#/components/schemas/SteamCurrency"
        amountRules:
          type: array
          items:
            $ref: "#/components/schemas/SteamAmountRule"
        minimumAmount:
          type: "null"
        maximumAmount:
          type: "null"
    SteamQuoteInput:
      type: object
      required: [steamLogin, currency, amount]
      properties:
        steamLogin:
          type: string
          minLength: 1
          maxLength: 100
        currency:
          $ref: "#/components/schemas/SteamCurrency"
        amount:
          type: string
          minLength: 1
          maxLength: 32
    SteamQuote:
      type: object
      required: [canRefill, currency, amount, priceRub]
      properties:
        canRefill:
          type: boolean
          enum: [true]
        currency:
          $ref: "#/components/schemas/SteamCurrency"
        amount:
          type: string
        priceRub:
          type: integer
          minimum: 1