# lib/api-zod/src/generated/types/checkoutConfig.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutField } from "./checkoutField";
import type { OrderType } from "./orderType";

export interface CheckoutConfig {
  orderType: OrderType;
  supported: boolean;
  /** @nullable */
  message: string | null;
  fields: CheckoutField[];
}