# lib/api-zod/src/generated/types/checkoutField.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutFieldType } from "./checkoutFieldType";

export interface CheckoutField {
  key: string;
  label: string;
  type: CheckoutFieldType;
}