# lib/api-zod/src/generated/types/createOrderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderInputCheckoutData } from "./createOrderInputCheckoutData";

export interface CreateOrderInput {
  productSlug: string;
  variantId: string;
  email: string;
  checkoutKey: string;
  checkoutData?: CreateOrderInputCheckoutData;
}