# lib/api-zod/src/generated/types/createOrderResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderResult {
  publicId: string;
  accessToken: string;
  paymentUrl: string;
}