# lib/api-zod/src/generated/types/orderStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  created: "created",
  payment_pending: "payment_pending",
  payment_confirmed: "payment_confirmed",
  supplier_processing: "supplier_processing",
  fulfilled: "fulfilled",
  email_sent: "email_sent",
  payment_failed: "payment_failed",
  supplier_failed: "supplier_failed",
  email_failed: "email_failed",
  failed: "failed",
  cancelled: "cancelled",
  refunded: "refunded",
} as const;