# lib/api-zod/src/generated/types/orderType.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  gift_card: "gift_card",
  steam_topup: "steam_topup",
  telegram_stars: "telegram_stars",
  telegram_premium: "telegram_premium",
  game_topup: "game_topup",
} as const;