# lib/api-zod/src/generated/types/storefrontProduct.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutConfig } from "./checkoutConfig";
import type { SteamFormConfig } from "./steamFormConfig";
import type { StorefrontCategoryId } from "./storefrontCategoryId";
import type { StorefrontOffer } from "./storefrontOffer";

export interface StorefrontProduct {
  slug: string;
  categoryId: StorefrontCategoryId;
  title: string;
  description: string;
  flag: string;
  /** @nullable */
  region: string | null;
  available: boolean;
  offers: StorefrontOffer[];
  steamForm: SteamFormConfig | null;
  checkout: CheckoutConfig;
}